% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_1-1_line.R
\name{Add_1_line}
\alias{Add_1_line}
\title{Adding 1-1 line to a ggplot}
\usage{
Add_1_line(
  type = "dashed",
  color = "black",
  size = 0.6,
  add_zero_line = FALSE,
  ...
)
}
\arguments{
\item{type}{linetype}

\item{color}{line color}

\item{size}{line size}

\item{add_zero_line}{logical, add 0-0 lines}

\item{...}{Other arguments passed to \code{geom_abline}}
}
\description{
Adding 1-1 line to a ggplot
}
\details{
This is a dummy workaround of the original \code{geom_abline}
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mtcars, aes(mpg, drat)) +
expand_limits(x = c(.0001, 0.0001),
              y = c(.0001, 0.0001)) +
  Add_1_line(add_zero_line = T) +
  geom_point() +
  theme_minimal()

}

}
