% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{Add_R2}
\alias{Add_R2}
\title{Adding R2 value for a linear regression}
\usage{
Add_R2(method = "lm", formula = "y ~ x", ...)
}
\arguments{
\item{method}{model type}

\item{formula}{model formula}

\item{...}{Other arguments passed to \code{stat_poly_eq}}
}
\description{
Adding R2 value for a linear regression
}
\details{
This is a dummy workaround of the original \code{stat_poly_eq}
}
\examples{
\dontrun{
library(ggpubr)

ggplot(mtcars, aes(wt, mpg)) +
    geom_point() +
    labs(title = "A Lovely Plot",
         subtitle = "What can the subtitle tell us?") +
         Add_R2() +
    theme_clean()
}

}
