% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atslib-package.R
\docType{package}
\name{atslib-package}
\alias{atslib}
\alias{atslib-package}
\title{atslib: Personal R package of Anatoly Tsyplenkov}
\description{
R functions for my personal use, mostly plotting!
}
\keyword{internal}
