% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_plot.R
\name{multi_plot}
\alias{multi_plot}
\title{Interactive graph of several Y-variables}
\usage{
multi_plot(.data, .x, .y, .c4a_pal = "tableau.color_blind", ...)
}
\arguments{
\item{.data}{dataframe.}

\item{.x}{character.}

\item{.y}{character.}

\item{.c4a_pal}{character.}

\item{...}{other params applied to dyOptions}
}
\value{
a dygraph plot
}
\description{
Interactive graph of several Y-variables
}
