% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_round.R
\name{smart_signif}
\alias{smart_signif}
\title{Rounding to 3 signif digits}
\usage{
smart_signif(x)
}
\arguments{
\item{x}{numerical value}
}
\description{
This function rounds decimal values to a 3 signif digits.
}
\examples{
smart_signif(3.4699878)
}
\keyword{round}
