% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{theme_clean}
\alias{theme_clean}
\title{Minimal ggplot2 theme using the Ubuntu Condensed and Ubuntu fonts}
\usage{
theme_clean(
  base_font_family = "Ubuntu",
  base_font_size = 12,
  legend = "bottom",
  ...
)
}
\arguments{
\item{base_font_family}{base font family}

\item{base_font_size}{base font size}

\item{legend}{legend position}

\item{...}{Other arguments passed to \code{theme_pubclean}}
}
\description{
Minimal ggplot2 theme using the Ubuntu Condensed and Ubuntu fonts
}
\details{
The Ubuntu Condensed and Ubuntu fonts are both Google fonts;
they can be found at \url{https://fonts.google.com/specimen/Ubuntu+Condensed}
and \url{https://fonts.google.com/specimen/Ubuntu}. These fonts must be
installed locally on your computer for this theme to work.
}
\examples{
\dontrun{
library(ggpubr)

ggplot(mtcars, aes(wt, mpg)) +
    geom_point() +
    labs(title = "A Lovely Plot",
         subtitle = "What can the subtitle tell us?") +
    theme_clean()
}

}
