% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{theme_map}
\alias{theme_map}
\title{A ggplot theme for maps}
\usage{
theme_map(...)
}
\arguments{
\item{...}{Other arguments passed to \code{theme_minimal}}
}
\description{
A ggplot theme for maps
}
\details{
This is a slightly modified Timo Grossenbacher theme from
this tutorial \url{https://timogrossenbacher.ch/2018/03/categorical-spatial-interpolation-with-r/}
}
