% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{Add_R2}
\alias{Add_R2}
\title{Adding R2 value for a linear regression}
\usage{
Add_R2(
  method = "lm",
  formula = "y ~ x",
  add_line = T,
  lty = "dashed",
  conf_int = F,
  ...
)
}
\arguments{
\item{method}{model type}

\item{formula}{model formula}

\item{add_line}{Should we add a \code{geom_smooth} to the plot?}

\item{lty}{linetype of the \code{geom_smooth}}

\item{conf_int}{Should we display confindence intervals of the \code{geom_smooth}?}

\item{...}{Other arguments passed to \code{stat_poly_eq}}

\item{color}{color of the \code{geom_smooth}}
}
\description{
Adding R2 value for a linear regression
}
\details{
This is a dummy workaround of the original \code{stat_poly_eq}
}
\examples{
\dontrun{
library(ggpubr)

ggplot(mtcars, aes(wt, mpg)) +
    geom_point() +
    labs(title = "A Lovely Plot",
         subtitle = "What can the subtitle tell us?") +
         Add_R2() +
    theme_clean()
}

}
