% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atslib-package.R
\docType{package}
\name{atslib-package}
\alias{atslib}
\alias{atslib-package}
\title{atslib: Personal R package of Anatoly Tsyplenkov}
\description{
R functions for my personal use, mostly plotting!
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atsyplenkov/atslib}
  \item Report bugs at \url{https://github.com/atsyplenkov/atslib/issues}
}

}
\keyword{internal}
